+++ platform/template/platform.c
@@ -156,5 +156,5 @@ const struct sbi_platform platform = {
 	.hart_count		= 1,
 	.hart_stack_size	= SBI_PLATFORM_DEFAULT_HART_STACK_SIZE,
 	.heap_size		= SBI_PLATFORM_DEFAULT_HEAP_SIZE(1),
-	.platform_ops_addr	= (unsigned long)&platform_ops
+	.platform_ops_addr	= (uintptr_t)&platform_ops
 };
